requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/Bogdanp/dramatiq.git#2c3c47cc356e3034db002555a9ddd216e73a12ce#egg=dramatiq
-e .
asn1crypto==0.24.0        # via cryptography
attrs==18.1.0             # via automat, service-identity
automat==0.6.0            # via twisted
beautifulsoup4==4.6.0
cffi==1.11.5              # via cryptography
colorlog==3.1.4
constantly==15.1.0        # via twisted
cryptography==2.2.2       # via pyopenssl
cssselect==1.0.3          # via parsel, scrapy
django-dramatiq==0.3.0
django-environ==0.4.4
django-fsm==2.6.0
django-model-utils==3.1.2
django==1.11.13           # via django-dramatiq, django-environ, django-model-utils, jsonfield2
djangorestframework==3.8.2
elasticsearch-dsl==6.1.0
elasticsearch==6.2.0      # via elasticsearch-dsl
hyperlink==18.0.0         # via twisted
idna==2.6                 # via cryptography, hyperlink
incremental==17.5.0       # via twisted
ipaddress==1.0.22         # via elasticsearch-dsl
jsonfield2==3.0.1
lxml==4.2.1               # via parsel, scrapy
nltk==3.3.0
parsel==1.4.0             # via scrapy
progressbar2==3.37.1
prometheus-client==0.0.20
pyasn1-modules==0.2.1     # via service-identity
pyasn1==0.4.3             # via pyasn1-modules, service-identity
pycparser==2.18           # via cffi
pydispatcher==2.0.5       # via scrapy
pyopenssl==18.0.0         # via scrapy, service-identity
python-dateutil==2.7.3    # via elasticsearch-dsl
python-utils==2.3.0       # via progressbar2
pytz==2018.4              # via django
queuelib==1.5.0           # via scrapy
raven==6.8.0
redis==2.10.6
scrapy==1.5.0
service-identity==17.0.0  # via scrapy
shortuuid==0.5.0
six==1.11.0               # via automat, cryptography, django-environ, elasticsearch-dsl, nltk, parsel, progressbar2, pyopenssl, python-dateutil, python-utils, scrapy, w3lib
terminaltables==3.1.0
twisted==18.4.0           # via scrapy
urllib3==1.22             # via elasticsearch
w3lib==1.19.0             # via parsel, scrapy
zope.interface==4.5.0     # via twisted
